requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
supabase==2.10.0
yfinance>=1.2.0
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.27.2
